feat(payments): lessons_count as DecimalField for fractional refund remainders
Refund rows will store a negative fractional lesson count (half-lesson
remainders). lessons_count becomes numeric(12,1); fifo_inputs keeps the exact
decimal for refund consumptions; API still returns lessons_count as a JS number
(_js_number) so whole-lesson output is byte-identical.

+++ journal_django/apps/payments/models.py
@@ -52,7 +52,7 @@ class Payment(models.Model):
         blank=True,
     )
     subscriptions_count = models.IntegerField(null=True, blank=True)
-    lessons_count = models.IntegerField(null=True, blank=True)
+    lessons_count = models.DecimalField(max_digits=12, decimal_places=1, null=True, blank=True)
     kind = models.TextField(default='purchase', db_default='purchase')
     unit_price = models.DecimalField(max_digits=10, decimal_places=2)
     total_amount = models.DecimalField(max_digits=10, decimal_places=2)
